+++ core/src/main/scala/portals/builder/FlowBuilderImpl.scala
@@ -139,7 +139,7 @@ class FlowBuilderImpl[T, U, CT, CU](using fbctx: FlowBuilderContext[T, U]) exten
     val behavior = Tasks.map[CU, CCU](f)
     addTask(behavior)
 
-  override def key(f: CU => Int): FlowBuilder[T, U, CU, CU] =
+  override def key(f: CU => Long): FlowBuilder[T, U, CU, CU] =
     val behavior = Tasks.key[CU](f)
     addTask(behavior)
