+++ core/src/main/scala/portals/system/WrappedEvent.scala
@@ -1,7 +1,7 @@
 package portals
 
 sealed trait WrappedEvent[+T]
-case class Event[T](key: Key[Int], event: T) extends WrappedEvent[T]
+case class Event[T](key: Key[Long], event: T) extends WrappedEvent[T]
 case class Error[T](t: Throwable) extends WrappedEvent[T]
 case object Atom extends WrappedEvent[Nothing]
 case object Seal extends WrappedEvent[Nothing]
@@ -14,14 +14,14 @@ private[portals] case class PortalMeta(
 )
 
 private[portals] case class Ask[T](
-    key: Key[Int],
+    key: Key[Long],
     meta: PortalMeta,
     event: T
 ) extends WrappedEvent[T]
 
 /** internal API */
 private[portals] case class Reply[T](
-    key: Key[Int],
+    key: Key[Long],
     meta: PortalMeta,
     event: T
 ) extends WrappedEvent[T]
