+++ core/src/main/scala/portals/tasks/TaskCallback.scala
@@ -5,8 +5,8 @@ trait TaskCallback[T, U, Req, Rep]:
   def submit(event: WrappedEvent[U]): Unit
 
   // Asker Task
-  def ask(portal: String, asker: String, req: Req, key: Key[Int], id: Int): Unit
+  def ask(portal: String, asker: String, req: Req, key: Key[Long], id: Int): Unit
 
   // Replier Task
-  def reply(r: Rep, portal: String, asker: String, key: Key[Int], id: Int): Unit
+  def reply(r: Rep, portal: String, asker: String, key: Key[Long], id: Int): Unit
 end TaskCallback // trait
